Fix Jenkinsfile
The Deploy stage now only gets executed if a environment variable "TOSCANA_DEPLOY_ON_BUILD" is set to true. otherwise this should be skipped. This change is necessary to Let the build pass even if you do not want to deploy. The current approach with the label is okay, but jenkins will not finish the build if there is no Node with the label "deploy". Because it waits for a worker to appear with this label.