Fix "Too many open Files" issue
The cause for this was a File appender that was created and opened for
every logger instance created by the LogImpl class.

Because of a missing close operation (stop) this was causing too many
open files error messages.

This issue should be resolved by this commit. The ExecutionTask and
other classes using the Log interface have to close the instance after
the job is done to close the common FileAppenders