Fix input form not showing up on platform select
Problem: If the view state was null and something was pushed to the csar
observable it automaticly routed to / on the creation of a empty
transformation for the list on the left side.

Solution: Add a view state for a open input form view. This also enables
highlighting for the platform list entry in the left pane.